app/requirements.txt
argon2-cffi==19.1.0
cffi==1.12.3
Django==2.2.4
django-cors-headers==3.1.0
django-debug-toolbar==2.0
django-rest-auth==0.9.5
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
Faker==2.0.0
numpy==1.17.0
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.0
python-memcached==1.59
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
django_heroku
gunicorn